refactor: move WinningCalculator package
